requirements.txt
asgiref==3.2.3
boto3==1.11.13
botocore==1.14.13
Django==3.0.1
django-storages==1.9.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
docutils==0.15.2
jmespath==0.9.4
mysqlclient==1.4.6
Pillow==7.0.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
uWSGI==2.0.18
